Trying to use better types for IDs and timestamps
Second thoughts about hash table replacement

Preparing to add atomic compatiblity (via Boost's atomic module)

Added GenSym() to Lua modules

+++ utils/Blob.h
@@ -122,7 +122,7 @@ namespace BlobXS {
 
 	//
 	struct BlobPimpl {
-		typedef unsigned long long storage_id;
+		typedef size_t storage_id;
 
 		// Helper functions
 		static storage_id BadID (void);
